Move App|WalletSettings to `settings` folder
