+++ src/renderer/components/app/index.ts
@@ -1,2 +1,1 @@
-export { AppSettings } from './AppSettings'
 export * from './AppUpdate'
