+++ src/renderer/components/settings/index.ts
@@ -1,2 +1,3 @@
 export { WalletSettings } from './WalletSettings'
 export { UnlockWalletSettings } from './UnlockWalletSettings'
+export { AppSettings } from './AppSettings'
